Updated to handle writing multiple files to folder
As discovered in issue
[12](https://github.com/bezoerb/grunt-critical/issues/12), when trying
to write multiple files to a folder, the write fails because it assumes
that the path is a file and not a folder.

I have updated the code to check first if the destination is a folder
and update accordingly. See [this
issue](http://stackoverflow.com/a/20417119/335567) on SO for more
information.